+++ BoxingGame/Assets/Prefabs/Option.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7230d78cb2072fc4cbcb6ccd4f46d1a8
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
