+++ BoxingGame/Assets/_Script/OptionManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f97d13c0510197b438fd0c4268e48a4a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
